Rename files. Restore working order to GitHub auth.
